+++ src/setupProxy.js
@@ -1,6 +1,5 @@
 const proxy = require('http-proxy-middleware')
 
 module.exports = function(app) {
-  app.use(proxy('/__', { target: 'http://localhost:4000/' }))
-  app.use(proxy('/functions', { target: 'http://localhost:4000/' }))
+  app.use(proxy('/api', { target: 'http://localhost:8080', pathRewrite: { '^/api/': '/' } }))
 }